Split equipService
